+++ src/gtkutil.h
@@ -85,7 +85,6 @@ void action_radio(GSimpleAction *action,
 	GVariant *parameter, gpointer user_data);
 
 int get_dpi(void);
-guint get_modifiers(GtkEventController *controller);
 
 G_DEFINE_AUTOPTR_CLEANUP_FUNC(cairo_t, cairo_destroy)
 G_DEFINE_AUTOPTR_CLEANUP_FUNC(cairo_surface_t, cairo_surface_destroy)